requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-datetime-widget==0.9.3
django-filer==1.7.0
django-hero-slider==1.3.1
django-hvad==1.8.0
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.11.0
django-polymorphic==2.1.2
django-select2==7.2.3
django-tempus-dominus==5.1.2.13
djangorestframework==3.11.0
drf-yasg==1.17.1
easy-thumbnails==2.7
flower==0.9.4
gunicorn==20.0.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
packaging==20.3
Pillow==7.1.1
psycopg2==2.8.5
pyparsing==2.4.7
pytz==2019.3
redis==3.5.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
